\name{EcoHydro2D-package}
\alias{EcoHydro2D-package}
\alias{EcoHydro2D}
\docType{package}
\title{
\packageTitle{EcoHydro2D}
}
\description{
\packageDescription{EcoHydro2D}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EcoHydro2D}
\packageIndices{EcoHydro2D}
This package brings together the functions related to the two dimensional ecohydrological modelling based on Vervoort and van der Zee (2012)
}
\author{
\packageAuthor{EcoHydro2D}

Maintainer: \packageMaintainer{EcoHydro2D}
}
\references{
Vervoort and van der Zee (2012)
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%%~~ documentation directory ~~
\keyword{ package }
\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
